requirements.txt
django<2
django-registration-redux